Worked on addresses and country, state and city dropdowns.

+++ Models/DatabaseModels/DeliveryModel.cs
@@ -2,12 +2,12 @@
 {
     public class DeliveryModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public required int UserId { get; set; }
+        public required long UserId { get; set; }
         public UserModel User { get; set; } = null!;
 
-        public required int OrderId { get; set; }
+        public required long OrderId { get; set; }
         public OrderModel Order { get; set; } = null!;
     }
 }
