+++ Models/DatabaseModels/OrderedDishesModel.cs
@@ -2,9 +2,9 @@
 {
     public class OrderedDishesModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public required int OrderId { get; set; }
+        public required long OrderId { get; set; }
         public OrderModel Order { get; set; } = null!;
 
         public required int DishId { get; set; }
